make zuul-cloner tesora-repo-mapping-aware
The idea is that we can handle the cloning of repos with
canonical names like 'openstack/trove', but know to apply the
correct upstream which is tesora's repo.

Before this we were failing to update the cached bits for some
tesora repositories - using the /opt/git cache that nodepool
created and applying the patch on top of that.  This could be
significantly out of date.

Change-Id: Ia51174725d6ac75dc0066f6a864213d956485e60